Minor armv7l g++ 8.3.0 build fixed
Used 32bit machine: Raspbian 10.4 'buster', armv7l (armhf), g++ 8.3.0, OpenJDK 11.0.7

Note that main development occurs on GNU/Linux Debian 11 'bullseye', amd64, g++ 9.3.0, OpenJDK 11.0.7

+++ api/direct_bt/dbt_debug.hpp
@@ -65,12 +65,12 @@ extern "C" {
 
 template<class ListElemType>
 inline void printSharedPtrList(std::string prefix, std::vector<std::shared_ptr<ListElemType>> & list) {
-    fprintf(stderr, "%s: Start: %zd elements\n", prefix.c_str(), list.size());
+    fprintf(stderr, "%s: Start: %zd elements\n", prefix.c_str(), (size_t)list.size());
     int idx = 0;
     for (auto it = list.begin(); it != list.end(); idx++) {
         std::shared_ptr<ListElemType> & e = *it;
         if ( nullptr != e ) {
-            fprintf(stderr, "%s[%d]: useCount %zd, mem %p\n", prefix.c_str(), idx, e.use_count(), e.get());
+            fprintf(stderr, "%s[%d]: useCount %zd, mem %p\n", prefix.c_str(), idx, (size_t)e.use_count(), e.get());
         } else {
             fprintf(stderr, "%s[%d]: NULL\n", prefix.c_str(), idx);
         }
